Salvando dados

+++ 7.Listas/5.exc_aprvd_rprvd_sem_def.py
@@ -5,7 +5,7 @@ QUANTIDADE_NOTAS = 4
 lista_notas = []
 
 for i in range(QUANTIDADE_NOTAS):
-    nota = float(input("Digite uma nota: "))
+    nota = int(input("Digite uma nota: "))
     lista_notas.append(nota)
 
 media = sum(lista_notas) / QUANTIDADE_NOTAS
